# client/docs/openapi.yaml
openapi: 3.1.0
info:
  title: BUZZZ Platform API
  version: 1.0.0
  description: 'The single backend interface for BUZZZ. Every endpoint is workspace scoped, permission checked and audited.
    Side-effecting endpoints accept an idempotency key.


    **Status: specification.** No endpoint below is implemented yet; the current product is a browser artifact with zero server
    routes. This document defines the target so implementation is consistent.'
  contact:
    name: BUZZZ
    url: https://buzzzbuzzz.com
servers:
- url: https://api.buzzzbuzzz.com/api/v1
  description: Production
- url: https://staging-api.buzzzbuzzz.com/api/v1
  description: Staging
security:
- bearerAuth: []
...
paths:
...
  /contacts:
    get:
      tags:
      - CRM
      summary: List Contacts
      operationId: listContacts
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Sort'
      - $ref: '#/components/parameters/Query'
      - $ref: '#/components/parameters/Filter'
      responses:
        '200':
          description: Page of Contacts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
      - CRM
      summary: Create a Contact
      operationId: createContact
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Contact'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Validation'
        '402':
          $ref: '#/components/responses/PlanLimit'
...
  /companies:
    get:
      tags:
      - CRM
      summary: List Companys
      operationId: listCompanys
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Sort'
      - $ref: '#/components/parameters/Query'
      - $ref: '#/components/parameters/Filter'
      responses:
        '200':
          description: Page of Companys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
      - CRM
      summary: Create a Company
      operationId: createCompany
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Company'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Validation'
        '402':
          $ref: '#/components/responses/PlanLimit'
...
  /deals:
    get:
      tags:
      - CRM
      summary: List Deals
      operationId: listDeals
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Sort'
      - $ref: '#/components/parameters/Query'
      - $ref: '#/components/parameters/Filter'
      responses:
        '200':
          description: Page of Deals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
      - CRM
      summary: Create a Deal
      operationId: createDeal
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Deal'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deal'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Validation'
        '402':
          $ref: '#/components/responses/PlanLimit'
...
  /tasks:
    get:
      tags:
      - CRM
      summary: List Tasks
      operationId: listTasks
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Sort'
      - $ref: '#/components/parameters/Query'
      - $ref: '#/components/parameters/Filter'
      responses:
        '200':
          description: Page of Tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
      - CRM
      summary: Create a Task
      operationId: createTask
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Validation'
        '402':
          $ref: '#/components/responses/PlanLimit'
...
  /conversations:
    get:
      tags:
      - Inbox
      summary: List Conversations
      operationId: listConversations
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Sort'
      - $ref: '#/components/parameters/Query'
      - $ref: '#/components/parameters/Filter'
      responses:
        '200':
          description: Page of Conversations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
      - Inbox
      summary: Create a Conversation
      operationId: createConversation
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Conversation'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Validation'
        '402':
          $ref: '#/components/responses/PlanLimit'
...
  /conversations/{id}/messages:
    get:
      tags:
      - Inbox
      summary: List messages in a conversation
      description: ''
      parameters:
      - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: OK
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
    post:
      tags:
      - Inbox
      summary: Send a message
      description: 'Runs the communication orchestrator: consent, preference, quiet hours, frequency and agent governance.'
      parameters:
      - $ref: '#/components/parameters/Id'
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: OK
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /appointments:
    get:
      tags:
      - Appointments
      summary: List Appointments
      operationId: listAppointments
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Sort'
      - $ref: '#/components/parameters/Query'
      - $ref: '#/components/parameters/Filter'
      responses:
        '200':
          description: Page of Appointments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
      - Appointments
      summary: Create a Appointment
      operationId: createAppointment
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Appointment'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Validation'
        '402':
          $ref: '#/components/responses/PlanLimit'
...
  /agents:
    get:
      tags:
      - Agents
      summary: List Agents
      operationId: listAgents
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Sort'
      - $ref: '#/components/parameters/Query'
      - $ref: '#/components/parameters/Filter'
      responses:
        '200':
          description: Page of Agents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
      - Agents
      summary: Create a Agent
      operationId: createAgent
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Agent'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Validation'
        '402':
          $ref: '#/components/responses/PlanLimit'
...
  /workflows:
    get:
      tags:
      - Automations
      summary: List Workflows
      operationId: listWorkflows
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Sort'
      - $ref: '#/components/parameters/Query'
      - $ref: '#/components/parameters/Filter'
      responses:
        '200':
          description: Page of Workflows
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
      - Automations
      summary: Create a Workflow
      operationId: createWorkflow
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Workflow'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Validation'
        '402':
          $ref: '#/components/responses/PlanLimit'
...
  /events:
    get:
      tags:
      - Events
      summary: Query the event stream
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Filter'
      responses:
        '200':
          description: Events
...
  /webhooks/gowhats:
    post:
      tags:
      - Webhooks
      summary: GoWhats inbound webhook
      description: Signature verified. Idempotent on the provider event id.
      security: []
      responses:
        '200':
          description: Accepted
        '400':
          description: Signature verification failed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short lived access token. The workspace is taken from the token, never from the request body.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: cursor
      in: query
      schema:
        type: string
      description: Keyset cursor. Page fifty costs the same as page one.
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        maximum: 200
    Sort:
      name: sort
      in: query
      schema:
        type: string
      description: Field name, prefix with - for descending.
    Query:
      name: q
      in: query
      schema:
        type: string
    Filter:
      name: filter
      in: query
      schema:
        type: string
      description: field:op:value, repeatable. Example filter=status:eq:Qualified
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: Required on anything that sends, charges or dials. A repeat returns the original result rather than acting
        twice.
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but not permitted. The body names the required role or autonomy level.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Conflict:
      description: Conflicts with current state, such as a double booking
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Validation:
      description: Validation failed; field errors included
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    PlanLimit:
      description: Blocked by the plan entitlement or usage allowance
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required:
      - code
      - message
      - request_id
      properties:
        code:
          type: string
          example: autonomy_insufficient
        message:
          type: string
          example: Create deals needs autonomy Level 3. Kai runs at Level 2.
        request_id:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
        docs:
          type: string
          format: uri
    Page:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        next_cursor:
          type:
          - string
          - 'null'
        has_more:
          type: boolean
    Contact:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        company:
          type: string
        title:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
        source:
          type: string
        owner:
          type: string
        stage:
          type: string
        location:
          type: string
        score:
          type: integer
          readOnly: true
          description: Computed on read from engagement, appointments, deals and memory.
        tags:
          type: array
          items:
            type: string
        channels:
          type: array
          items:
            type: string
        workspace_id:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
    Company:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        domain:
          type: string
    Deal:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        contact_id:
          type: string
        value:
          type: integer
          description: Minor units
        stage:
          type: string
        probability:
          type: integer
        owner:
          type: string
...
    Task:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        due:
          type: string
          format: date-time
        assignee:
          type: string
        done:
          type: boolean
...
    Conversation:
      type: object
      properties:
        id:
          type: string
        contact_id:
          type: string
        channel:
          type: string
        state:
          type: string
          enum:
          - Open
          - Waiting
          - Escalated
          - Resolved
        assignee:
          type: string
        ai:
          type: boolean
...
    Appointment:
      type: object
      properties:
        id:
          type: string
        contact_id:
          type: string
        start:
          type: string
          format: date-time
        duration_minutes:
          type: integer
        staff_id:
          type: string
        service_id:
          type: string
        status:
          type: string
          enum:
          - Pending
          - Confirmed
          - In progress
          - Completed
          - Cancelled
          - No show
    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        title:
          type: string
        type:
          type: string
        status:
          type: string
          enum:
          - Draft
          - Active
          - Paused
          - Archived
        autonomy:
          type: integer
          minimum: 0
          maximum: 4
          description: Capped at run time by the workspace ceiling.
        tools:
          type: array
          items:
            type: string
        permissions:
          type: object
          additionalProperties:
            type: boolean
        guardrails:
          type: array
          items:
            type: string
        escalation:
          type: object
...
    Workflow:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum:
          - Draft
          - Published
          - Paused
          - Archived
        version:
          type: integer
        graph:
          type: object
          properties:
            nodes:
              type: array
              items:
                type: object
            edges:
              type: array
              items:
                type: object
...